requirements.txt
addict==2.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
fonttools==4.34.4
idna==3.3
importlib-metadata==4.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.5.2
mmcv-full==1.6.1
mmdet==2.25.1
model-index==0.1.11
numpy==1.23.1
opencv-python==4.6.0.66
openmim==0.2.0
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pycocotools==2.0.4
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
rich==12.5.1
six==1.16.0
tabulate==0.8.10
terminaltables==3.1.10
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.11
yapf==0.32.0
zipp==3.8.1
